The version option outputs only the jiren version
The package version string is now automatically generated from git
tags and commits when the package is built.

As a result, the import-metadata package to get the version string is
no longer needed. It was also used to get the version strings of other
dependent packages, but this process has been dropped.

Before:
```
$ jiren --version
jiren, version 0.7.3
jinja2, version 3.1.2
```

After:
```
$ jiren --version
0.7.3
```